doctor_customer/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.1
Django==4.2.6
idna==3.4
Pillow==10.0.1
psycopg2-binary==2.9.9
pypng==0.20220715.0
python-decouple==3.8
python-environ==0.4.54
qrcode==7.4.2
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.7
